phase2_multiview/requirements.txt
numpy>=1.21.0
opencv-python>=4.5.0
streamlit>=1.20.0
Pillow>=9.0.0
